nvfp4: store global scales as bf16 to match model dtype
Previously gs_K_*, gs_V_* were saved as fp32. Switched to bf16 so the
on-disk global scale matches the model's bfloat16 working dtype. The
kernel (vllm-exaone-fork: kv_fake_quant/kernels.py) upcasts to fp32
internally for arithmetic so precision is preserved end-to-end —
the change is purely about storage / dtype alignment.

amax_* fields kept as fp32 (intermediate stats, not consumed by the
kernel; useful at full fp32 for diagnostic comparisons).

Note: existing logs/calib/nvfp4_global_scales_*.pt files were derived
under the old behaviour (fp32 storage, slightly different FP4 round
ties). Need regen+rerun across the NVFP4 sweep to bake in both fixes
together.

+++ scripts/derive_nvfp4_global_scales.py
@@ -104,14 +104,16 @@ def derive(calib: dict) -> dict:
     gs_V_smooth = NVFP4_NUMERATOR / amax_V_smooth.clamp(min=1e-30)
 
     return {
-        # SmoothKV+NVFP4 (input to NVFP4 = K/s_K)
-        'gs_K_smooth':    gs_K_smooth.float(),
-        'gs_V_smooth':    gs_V_smooth.float(),
+        # SmoothKV+NVFP4 (input to NVFP4 = K/s_K).
+        # Stored as bf16 to match model dtype (kernels.py upcasts to fp32 for
+        # the per-element scale arithmetic, so precision is preserved).
+        'gs_K_smooth':    gs_K_smooth.bfloat16(),
+        'gs_V_smooth':    gs_V_smooth.bfloat16(),
         'amax_K_smooth':  amax_K_smooth.float(),
         'amax_V_smooth':  amax_V_smooth.float(),
         # NVFP4-only (input = K_raw)
-        'gs_K_raw':       gs_K_raw.float(),
-        'gs_V_raw':       gs_V_raw.float(),
+        'gs_K_raw':       gs_K_raw.bfloat16(),
+        'gs_V_raw':       gs_V_raw.bfloat16(),
         'amax_K_raw':     amax_K_raw.float(),
         'amax_V_raw':     amax_V_raw.float(),
         # Bookkeeping
